refactor: resolve-dependendencies
