+++ pkg-manager/resolve-dependencies/src/resolvePeers.ts
@@ -8,22 +8,19 @@ import type {
   PeerDependencyIssuesByProjects,
 } from '@pnpm/types'
 import { depPathToFilename, createPeersFolderSuffix } from '@pnpm/dependency-path'
-import normalize from 'normalize-path'
 import mapValues from 'ramda/src/map'
 import partition from 'ramda/src/partition'
 import pick from 'ramda/src/pick'
 import scan from 'ramda/src/scan'
 import {
-  type LinkedDependency,
   type DependenciesTree,
   type DependenciesTreeNode,
   type ResolvedPackage,
 } from './resolveDependencies'
-import {
-  type ResolvedImporters,
-} from './resolveDependencyTree'
+import { type ResolvedImporters } from './resolveDependencyTree'
 import { mergePeers } from './mergePeers'
 import { createNodeId, splitNodeId } from './nodeIdUtils'
+import { dedupeInjectedDeps } from './dedupeInjectedDeps'
 
 export interface GenericDependenciesGraphNode {
   // at this point the version is really needed only for logging
@@ -142,64 +139,6 @@ export function resolvePeers<T extends PartialResolvedPackage> (
     dependenciesGraph: depGraph,
     dependenciesByProjectId,
     peerDependencyIssuesByProjects,
-  }
-}
-
-function dedupeInjectedDeps<T extends PartialResolvedPackage> (
-  opts: {
-    depGraph: GenericDependenciesGraph<T>
-    dependenciesByProjectId: DependenciesByProjectId
-    lockfileDir: string
-    pathsByNodeId: Map<string, string>
-    projects: ProjectToResolve[]
-    resolvedImporters: ResolvedImporters
-  }
-) {
-  const injectedDepsByProjects = new Map<string, Map<string, { depPath: string, id: string }>>()
-  for (const project of opts.projects) {
-    for (const [alias, nodeId] of Object.entries(project.directNodeIdsByAlias)) {
-      const depPath = opts.pathsByNodeId.get(nodeId)!
-      if (!depPath.startsWith('file:')) continue
-      const id = opts.depGraph[depPath].id.substring(5)
-      if (opts.projects.some((project) => project.id === id)) {
-        if (!injectedDepsByProjects.has(project.id)) injectedDepsByProjects.set(project.id, new Map())
-        injectedDepsByProjects.get(project.id)!.set(alias, { depPath, id })
-      }
-    }
-  }
-  const toDedupe = new Map<string, Map<string, string>>()
-  for (const [id, deps] of injectedDepsByProjects.entries()) {
-    const dedupedInjectedDeps = new Map<string, string>()
-    for (const [alias, dep] of deps.entries()) {
-      // Check for subgroup not equal.
-      // The injected project in the workspace may have dev deps
-      const isSubset = Object.entries(opts.depGraph[dep.depPath].children)
-        .every(([alias, depPath]) => opts.dependenciesByProjectId[dep.id][alias] === depPath)
-      if (isSubset) {
-        dedupedInjectedDeps.set(alias, dep.id)
-      }
-    }
-    toDedupe.set(id, dedupedInjectedDeps)
-  }
-  for (const [id, aliases] of toDedupe.entries()) {
-    for (const [alias, dedupedProjectId] of aliases.entries()) {
-      delete opts.dependenciesByProjectId[id][alias]
-      const index = opts.resolvedImporters[id].directDependencies.findIndex((dep) => dep.alias === alias)
-      const prev = opts.resolvedImporters[id].directDependencies[index]
-      const depPath = `link:${normalize(path.relative(id, dedupedProjectId))}`
-      const linkedDep: LinkedDependency = {
-        ...prev,
-        isLinkedDependency: true,
-        depPath,
-        pkgId: depPath,
-        resolution: {
-          type: 'directory',
-          directory: path.join(opts.lockfileDir, dedupedProjectId),
-        },
-      }
-      opts.resolvedImporters[id].directDependencies[index] = linkedDep
-      opts.resolvedImporters[id].linkedDependencies.push(linkedDep)
-    }
   }
 }
